requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
google-auth==2.35.0
pyotp==2.9.0
pydantic==2.10.3
pydantic[email]==2.10.3
email-validator==2.2.0
jinja2==3.1.4
requests==2.32.3
httpx==0.28.1